refactor: 优化代码 1. 更改`NmsRecipeRegister`接口名字为`CraftorithmRecipeRegister` 2. 去除用于测试的test物品源 3. `config.yml`中`use_nms_recipe_register`更改为`use_experimental_recipe_register` 4. 酿造台配方开始使用ItemIdRecipeChoice进行材料匹配 5. 优化BukkitRecipeChoiceParser中的报错提示
