+++ core/src/main/java/pers/yufiria/craftorithm/Craftorithm.java
@@ -22,7 +22,6 @@ import pers.yufiria.craftorithm.recipe.RecipeManager;
 import pers.yufiria.craftorithm.recipe.RecipeType;
 import pers.yufiria.craftorithm.script.ActionModule;
 import pers.yufiria.craftorithm.script.ConditionModule;
-import pers.yufiria.craftorithm.test.TestHook;
 import pers.yufiria.craftorithm.util.LangUtils;
 import pers.yufiria.craftorithm.util.UpdateChecker;
 
@@ -55,7 +54,6 @@ public final class Craftorithm extends BukkitPlugin implements LifeCycleTask {
 
     @Override
     public void whenEnable() {
-        api().registerItemPluginHook(TestHook.INSTANCE);
         ScriptEngine.INSTANCE.registerModule(ActionModule.INSTANCE);
         ScriptEngine.INSTANCE.registerModule(ConditionModule.INSTANCE);
         UpdateChecker.pullUpdateCheckRequest(Bukkit.getConsoleSender());
