+++ core/src/main/java/pers/yufiria/craftorithm/recipe/RecipeRegister.java
@@ -4,7 +4,6 @@ import org.bukkit.NamespacedKey;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.inventory.Recipe;
 import org.jetbrains.annotations.Nullable;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
 
 public interface RecipeRegister {
