+++ core/src/main/java/pers/yufiria/craftorithm/recipe/choice/BrewingRecipeChoiceParser.java
@@ -1,0 +1,18 @@
+package pers.yufiria.craftorithm.recipe.choice;
+
+import io.papermc.paper.potion.PotionMix;
+import org.bukkit.inventory.RecipeChoice;
+
+public enum BrewingRecipeChoiceParser implements RecipeChoiceParser {
+
+    INSTANCE;
+
+    @Override
+    public RecipeChoice parse(String choiceStr) {
+        RecipeChoice itemIdChoice = ItemIdRecipeChoiceParser.INSTANCE.parse(choiceStr);
+        return PotionMix.createPredicateChoice(
+            itemIdChoice
+        );
+    }
+
+}
