+++ core/src/main/java/pers/yufiria/craftorithm/ui/creator/anvil/AnvilCreator.java
@@ -20,7 +20,6 @@ import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import pers.yufiria.craftorithm.Craftorithm;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
 import pers.yufiria.craftorithm.config.Languages;
 import pers.yufiria.craftorithm.config.menu.creator.AnvilCreatorConfig;
 import pers.yufiria.craftorithm.item.ItemManager;
