+++ core/src/main/java/pers/yufiria/craftorithm/ui/creator/vanillaCrafting/VanillaShapedCreator.java
@@ -13,7 +13,6 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.recipe.CraftingBookCategory;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
 import pers.yufiria.craftorithm.config.Languages;
 import pers.yufiria.craftorithm.config.menu.creator.VanillaShapedCreatorConfig;
 import pers.yufiria.craftorithm.item.ItemManager;
