+++ core/src/main/java/pers/yufiria/craftorithm/ui/creator/vanillaSmithing/VanillaSmithingTransformCreator.java
@@ -12,7 +12,6 @@ import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
 import pers.yufiria.craftorithm.config.Languages;
 import pers.yufiria.craftorithm.config.menu.creator.VanillaSmithingTransformCreatorConfig;
 import pers.yufiria.craftorithm.item.ItemManager;
