+++ nms/v1_20/src/main/java/pers/yufiria/craftorithm/recipe/nms/CraftorithmRecipeRegister12000.java
@@ -1,0 +1,67 @@
+package pers.yufiria.craftorithm.recipe.nms;
+
+import crypticlib.MinecraftVersion;
+import crypticlib.lifecycle.LifeCycle;
+import crypticlib.lifecycle.LifeCycleTask;
+import crypticlib.lifecycle.LifeCycleTaskSettings;
+import crypticlib.lifecycle.TaskRule;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.world.item.crafting.IRecipe;
+import org.bukkit.NamespacedKey;
+import org.bukkit.inventory.*;
+import pers.yufiria.craftorithm.api.recipe.CraftorithmRecipeRegister;
+import pers.yufiria.craftorithm.recipe.RecipeManager;
+
+@LifeCycleTaskSettings(rules = {
+    @TaskRule(lifeCycle = LifeCycle.LOAD)
+})
+public enum CraftorithmRecipeRegister12000 implements CraftorithmRecipeRegister, LifeCycleTask {
+
+    INSTANCE;
+
+    @Override
+    public RegisterResult registerRecipe(Recipe bukkitRecipe) {
+        NamespacedKey recipeKey = RecipeManager.INSTANCE.getRecipeKey(bukkitRecipe);
+        IRecipe<?> nms;
+        switch (bukkitRecipe) {
+            case ShapedRecipe shapedRecipe -> {
+                nms = ShapedRecipe12000.fromBukkit(recipeKey, shapedRecipe);
+            }
+            case ShapelessRecipe shapelessRecipe -> {
+                nms = ShapelessRecipe12000.fromBukkit(recipeKey, shapelessRecipe);
+            }
+            case FurnaceRecipe furnaceRecipe -> {
+                nms = FurnaceRecipe12000.fromBukkit(recipeKey, furnaceRecipe);
+            }
+            case SmokingRecipe smokingRecipe -> {
+                nms = SmokingRecipe12000.fromBukkit(recipeKey, smokingRecipe);
+            }
+            case CampfireRecipe campfireRecipe -> {
+                nms = CampfireRecipe12000.fromBukkit(recipeKey, campfireRecipe);
+            }
+            case BlastingRecipe blastingRecipe -> {
+                nms = BlastingRecipe12000.fromBukkit(recipeKey, blastingRecipe);
+            }
+            case SmithingTransformRecipe smithingTransformRecipe -> {
+                nms = SmithingTransformRecipe12000.fromBukkit(recipeKey, smithingTransformRecipe);
+            }
+            case SmithingTrimRecipe smithingTrimRecipe -> {
+                nms = SmithingTrimRecipe12000.fromBukkit(recipeKey, smithingTrimRecipe);
+            }
+            case StonecuttingRecipe stonecuttingRecipe -> {
+                //切石机配方自定义匹配逻辑存在问题，暂时还是按照原版的来
+                nms = StonecuttingRecipe12000.fromBukkit(recipeKey, stonecuttingRecipe);
+            }
+            default -> {
+                return RegisterResult.UNSUPPORTED_RECIPE_TYPE;
+            }
+        }
+        MinecraftServer.getServer().aE().addRecipe(nms);
+        return RegisterResult.SUCCESS;
+    }
+
+    @Override
+    public void lifecycle(Object o, LifeCycle lifeCycle) {
+        REGISTER_COMPAT.register(MinecraftVersion.V1_20.name(), () -> this);
+    }
+}
