+++ nms/v1_20_2/src/main/java/pers/yufiria/craftorithm/recipe/nms/CraftorithmRecipeRegister12002.java
@@ -7,52 +7,51 @@ import crypticlib.lifecycle.LifeCycleTaskSettings;
 import crypticlib.lifecycle.TaskRule;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.item.crafting.IRecipe;
+import net.minecraft.world.item.crafting.RecipeHolder;
 import org.bukkit.NamespacedKey;
-import org.bukkit.craftbukkit.v1_20_R1.inventory.CraftStonecuttingRecipe;
 import org.bukkit.inventory.*;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
+import pers.yufiria.craftorithm.api.recipe.CraftorithmRecipeRegister;
 import pers.yufiria.craftorithm.recipe.RecipeManager;
-import pers.yufiria.craftorithm.util.RecipeUtils;
 
 @LifeCycleTaskSettings(rules = {
     @TaskRule(lifeCycle = LifeCycle.LOAD)
 })
-public enum NmsRecipeRegister12000 implements NmsRecipeRegister, LifeCycleTask {
+public enum CraftorithmRecipeRegister12002 implements CraftorithmRecipeRegister, LifeCycleTask {
 
     INSTANCE;
 
     @Override
     public RegisterResult registerRecipe(Recipe bukkitRecipe) {
         NamespacedKey recipeKey = RecipeManager.INSTANCE.getRecipeKey(bukkitRecipe);
-        IRecipe<?> nms;
+        RecipeHolder<? extends IRecipe<?>> nms;
         switch (bukkitRecipe) {
             case ShapedRecipe shapedRecipe -> {
-                nms = ShapedRecipe12000.fromBukkit(recipeKey, shapedRecipe);
+                nms = ShapedRecipe12002.fromBukkit(recipeKey, shapedRecipe);
             }
             case ShapelessRecipe shapelessRecipe -> {
-                nms = ShapelessRecipe12000.fromBukkit(recipeKey, shapelessRecipe);
+                nms = ShapelessRecipe12002.fromBukkit(recipeKey, shapelessRecipe);
             }
             case FurnaceRecipe furnaceRecipe -> {
-                nms = FurnaceRecipe12000.fromBukkit(recipeKey, furnaceRecipe);
+                nms = FurnaceRecipe12002.fromBukkit(recipeKey, furnaceRecipe);
             }
             case SmokingRecipe smokingRecipe -> {
-                nms = SmokingRecipe12000.fromBukkit(recipeKey, smokingRecipe);
+                nms = SmokingRecipe12002.fromBukkit(recipeKey, smokingRecipe);
             }
             case CampfireRecipe campfireRecipe -> {
-                nms = CampfireRecipe12000.fromBukkit(recipeKey, campfireRecipe);
+                nms = CampfireRecipe12002.fromBukkit(recipeKey, campfireRecipe);
             }
             case BlastingRecipe blastingRecipe -> {
-                nms = BlastingRecipe12000.fromBukkit(recipeKey, blastingRecipe);
+                nms = BlastingRecipe12002.fromBukkit(recipeKey, blastingRecipe);
             }
             case SmithingTransformRecipe smithingTransformRecipe -> {
-                nms = SmithingTransformRecipe12000.fromBukkit(recipeKey, smithingTransformRecipe);
+                nms = SmithingTransformRecipe12002.fromBukkit(recipeKey, smithingTransformRecipe);
             }
             case SmithingTrimRecipe smithingTrimRecipe -> {
-                nms = SmithingTrimRecipe12000.fromBukkit(recipeKey, smithingTrimRecipe);
+                nms = SmithingTrimRecipe12002.fromBukkit(recipeKey, smithingTrimRecipe);
             }
             case StonecuttingRecipe stonecuttingRecipe -> {
                 //切石机配方自定义匹配逻辑存在问题，暂时还是按照原版的来
-                nms = StonecuttingRecipe12000.fromBukkit(recipeKey, stonecuttingRecipe);
+                nms = StonecuttingRecipe12002.fromBukkit(recipeKey, stonecuttingRecipe);
             }
             default -> {
                 return RegisterResult.UNSUPPORTED_RECIPE_TYPE;
@@ -64,6 +63,6 @@ public enum NmsRecipeRegister12000 implements NmsRecipeRegister, LifeCycleTask {
 
     @Override
     public void lifecycle(Object o, LifeCycle lifeCycle) {
-        NMS_REGISTER_COMPAT.register(MinecraftVersion.V1_20.name(), () -> this);
+        REGISTER_COMPAT.register(MinecraftVersion.V1_20_2.name(), () -> this);
     }
 }
