+++ nms/v1_20_3/src/main/java/pers/yufiria/craftorithm/recipe/nms/CraftorithmRecipeRegister12003.java
@@ -10,13 +10,13 @@ import net.minecraft.world.item.crafting.IRecipe;
 import net.minecraft.world.item.crafting.RecipeHolder;
 import org.bukkit.NamespacedKey;
 import org.bukkit.inventory.*;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
+import pers.yufiria.craftorithm.api.recipe.CraftorithmRecipeRegister;
 import pers.yufiria.craftorithm.recipe.RecipeManager;
 
 @LifeCycleTaskSettings(rules = {
     @TaskRule(lifeCycle = LifeCycle.LOAD)
 })
-public enum NmsRecipeRegister12002 implements NmsRecipeRegister, LifeCycleTask {
+public enum CraftorithmRecipeRegister12003 implements CraftorithmRecipeRegister, LifeCycleTask {
 
     INSTANCE;
 
@@ -26,43 +26,42 @@ public enum NmsRecipeRegister12002 implements NmsRecipeRegister, LifeCycleTask {
         RecipeHolder<? extends IRecipe<?>> nms;
         switch (bukkitRecipe) {
             case ShapedRecipe shapedRecipe -> {
-                nms = ShapedRecipe12002.fromBukkit(recipeKey, shapedRecipe);
+                nms = ShapedRecipe12003.fromBukkit(recipeKey, shapedRecipe);
             }
             case ShapelessRecipe shapelessRecipe -> {
-                nms = ShapelessRecipe12002.fromBukkit(recipeKey, shapelessRecipe);
+                nms = ShapelessRecipe12003.fromBukkit(recipeKey, shapelessRecipe);
             }
             case FurnaceRecipe furnaceRecipe -> {
-                nms = FurnaceRecipe12002.fromBukkit(recipeKey, furnaceRecipe);
+                nms = FurnaceRecipe12003.fromBukkit(recipeKey, furnaceRecipe);
             }
             case SmokingRecipe smokingRecipe -> {
-                nms = SmokingRecipe12002.fromBukkit(recipeKey, smokingRecipe);
+                nms = SmokingRecipe12003.fromBukkit(recipeKey, smokingRecipe);
             }
             case CampfireRecipe campfireRecipe -> {
-                nms = CampfireRecipe12002.fromBukkit(recipeKey, campfireRecipe);
+                nms = CampfireRecipe12003.fromBukkit(recipeKey, campfireRecipe);
             }
             case BlastingRecipe blastingRecipe -> {
-                nms = BlastingRecipe12002.fromBukkit(recipeKey, blastingRecipe);
+                nms = BlastingRecipe12003.fromBukkit(recipeKey, blastingRecipe);
             }
             case SmithingTransformRecipe smithingTransformRecipe -> {
-                nms = SmithingTransformRecipe12002.fromBukkit(recipeKey, smithingTransformRecipe);
+                nms = SmithingTransformRecipe12003.fromBukkit(recipeKey, smithingTransformRecipe);
             }
             case SmithingTrimRecipe smithingTrimRecipe -> {
-                nms = SmithingTrimRecipe12002.fromBukkit(recipeKey, smithingTrimRecipe);
+                nms = SmithingTrimRecipe12003.fromBukkit(recipeKey, smithingTrimRecipe);
             }
             case StonecuttingRecipe stonecuttingRecipe -> {
-                //切石机配方自定义匹配逻辑存在问题，暂时还是按照原版的来
-                nms = StonecuttingRecipe12002.fromBukkit(recipeKey, stonecuttingRecipe);
+                nms = StonecuttingRecipe12003.fromBukkit(recipeKey, stonecuttingRecipe);
             }
             default -> {
                 return RegisterResult.UNSUPPORTED_RECIPE_TYPE;
             }
         }
-        MinecraftServer.getServer().aE().addRecipe(nms);
+        MinecraftServer.getServer().aG().addRecipe(nms);
         return RegisterResult.SUCCESS;
     }
 
     @Override
     public void lifecycle(Object o, LifeCycle lifeCycle) {
-        NMS_REGISTER_COMPAT.register(MinecraftVersion.V1_20_2.name(), () -> this);
+        REGISTER_COMPAT.register(MinecraftVersion.V1_20_3.name(), () -> this);
     }
 }
