+++ nms/v1_21_10_paper/src/main/java/pers/yufiria/craftorithm/recipe/nms/paper/CraftorithmRecipeRegister12110.java
@@ -1,5 +1,6 @@
-package pers.yufiria.craftorithm.recipe.nms;
+package pers.yufiria.craftorithm.recipe.nms.paper;
 
+import crypticlib.CrypticLibBukkit;
 import crypticlib.MinecraftVersion;
 import crypticlib.lifecycle.LifeCycle;
 import crypticlib.lifecycle.LifeCycleTask;
@@ -9,51 +10,55 @@ import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.item.crafting.*;
 import org.bukkit.Bukkit;
 import org.bukkit.NamespacedKey;
-import org.bukkit.craftbukkit.v1_21_R4.inventory.CraftStonecuttingRecipe;
+import org.bukkit.craftbukkit.inventory.CraftStonecuttingRecipe;
 import org.bukkit.inventory.*;
-import org.bukkit.inventory.FurnaceRecipe;
+import org.bukkit.inventory.BlastingRecipe;
+import org.bukkit.inventory.Recipe;
+import org.bukkit.inventory.ShapedRecipe;
+import org.bukkit.inventory.ShapelessRecipe;
 import org.bukkit.inventory.SmithingTransformRecipe;
 import org.bukkit.inventory.SmithingTrimRecipe;
+import org.bukkit.inventory.SmokingRecipe;
 import org.bukkit.inventory.TransmuteRecipe;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
+import pers.yufiria.craftorithm.api.recipe.CraftorithmRecipeRegister;
 import pers.yufiria.craftorithm.recipe.RecipeManager;
 import pers.yufiria.craftorithm.util.RecipeUtils;
 
 @LifeCycleTaskSettings(rules = {
     @TaskRule(lifeCycle = LifeCycle.LOAD)
 })
-public enum NmsRecipeRegister12105 implements NmsRecipeRegister, LifeCycleTask {
+public enum CraftorithmRecipeRegister12110 implements CraftorithmRecipeRegister, LifeCycleTask {
 
     INSTANCE;
 
     @Override
     public RegisterResult registerRecipe(Recipe bukkitRecipe) {
         NamespacedKey recipeKey = RecipeManager.INSTANCE.getRecipeKey(bukkitRecipe);
-        RecipeHolder<? extends IRecipe<?>> recipeHolder;
+        RecipeHolder<?> recipeHolder;
         switch (bukkitRecipe) {
             case ShapedRecipe shapedRecipe -> {
-                recipeHolder = ShapedRecipe12105.fromBukkit(recipeKey, shapedRecipe);
+                recipeHolder = ShapedRecipe12110.fromBukkit(recipeKey, shapedRecipe);
             }
             case ShapelessRecipe shapelessRecipe -> {
-                recipeHolder = ShapelessRecipe12105.fromBukkit(recipeKey, shapelessRecipe);
+                recipeHolder = ShapelessRecipe12110.fromBukkit(recipeKey, shapelessRecipe);
             }
             case FurnaceRecipe furnaceRecipe -> {
-                recipeHolder = FurnaceRecipe12105.fromBukkit(recipeKey, furnaceRecipe);
+                recipeHolder = SmeltingRecipe12110.fromBukkit(recipeKey, furnaceRecipe);
             }
             case SmokingRecipe smokingRecipe -> {
-                recipeHolder = SmokingRecipe12105.fromBukkit(recipeKey, smokingRecipe);
+                recipeHolder = SmokingRecipe12110.fromBukkit(recipeKey, smokingRecipe);
             }
             case CampfireRecipe campfireRecipe -> {
-                recipeHolder = CampfireRecipe12105.fromBukkit(recipeKey, campfireRecipe);
+                recipeHolder = CampfireCookingRecipe12110.fromBukkit(recipeKey, campfireRecipe);
             }
             case BlastingRecipe blastingRecipe -> {
-                recipeHolder = BlastingRecipe12105.fromBukkit(recipeKey, blastingRecipe);
+                recipeHolder = BlastingRecipe12110.fromBukkit(recipeKey, blastingRecipe);
             }
             case SmithingTransformRecipe smithingTransformRecipe -> {
-                recipeHolder = SmithingTransformRecipe12105.fromBukkit(recipeKey, smithingTransformRecipe);
+                recipeHolder = SmithingTransformRecipe12110.fromBukkit(recipeKey, smithingTransformRecipe);
             }
             case SmithingTrimRecipe smithingTrimRecipe -> {
-                recipeHolder = SmithingTrimRecipe12105.fromBukkit(recipeKey, smithingTrimRecipe);
+                recipeHolder = SmithingTrimRecipe12110.fromBukkit(recipeKey, smithingTrimRecipe);
             }
             case StonecuttingRecipe stonecuttingRecipe -> {
                 stonecuttingRecipe.setInputChoice(RecipeUtils.getBukkitChoice(stonecuttingRecipe.getInputChoice()));
@@ -68,13 +73,15 @@ public enum NmsRecipeRegister12105 implements NmsRecipeRegister, LifeCycleTask {
                 return RegisterResult.UNSUPPORTED_RECIPE_TYPE;
             }
         }
-        MinecraftServer.getServer().aI().addRecipe(recipeHolder);
+        MinecraftServer.getServer().getRecipeManager().addRecipe(recipeHolder);
         return RegisterResult.SUCCESS;
     }
 
     @Override
     public void lifecycle(Object plugin, LifeCycle lifeCycle) {
-        NMS_REGISTER_COMPAT.register(MinecraftVersion.V1_21_5.name(), () -> this);
+        //这个版本兼容1.21.9-1.21.11
+        if (CrypticLibBukkit.isPaper()) {
+            REGISTER_COMPAT.register(MinecraftVersion.V1_21_9.name(), () -> this);
+        }
     }
 }
-
