+++ nms/v1_21_10_spigot/src/main/java/pers/yufiria/craftorithm/recipe/nms/spigot/CraftorithmRecipeRegister12110.java
@@ -10,55 +10,51 @@ import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.item.crafting.*;
 import org.bukkit.Bukkit;
 import org.bukkit.NamespacedKey;
-import org.bukkit.craftbukkit.inventory.CraftStonecuttingRecipe;
+import org.bukkit.craftbukkit.v1_21_R6.inventory.CraftStonecuttingRecipe;
 import org.bukkit.inventory.*;
-import org.bukkit.inventory.BlastingRecipe;
-import org.bukkit.inventory.Recipe;
-import org.bukkit.inventory.ShapedRecipe;
-import org.bukkit.inventory.ShapelessRecipe;
+import org.bukkit.inventory.FurnaceRecipe;
 import org.bukkit.inventory.SmithingTransformRecipe;
 import org.bukkit.inventory.SmithingTrimRecipe;
-import org.bukkit.inventory.SmokingRecipe;
 import org.bukkit.inventory.TransmuteRecipe;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
+import pers.yufiria.craftorithm.api.recipe.CraftorithmRecipeRegister;
 import pers.yufiria.craftorithm.recipe.RecipeManager;
 import pers.yufiria.craftorithm.util.RecipeUtils;
 
 @LifeCycleTaskSettings(rules = {
     @TaskRule(lifeCycle = LifeCycle.LOAD)
 })
-public enum NmsRecipeRegister260100 implements NmsRecipeRegister, LifeCycleTask {
+public enum CraftorithmRecipeRegister12110 implements CraftorithmRecipeRegister, LifeCycleTask {
 
     INSTANCE;
 
     @Override
     public RegisterResult registerRecipe(Recipe bukkitRecipe) {
         NamespacedKey recipeKey = RecipeManager.INSTANCE.getRecipeKey(bukkitRecipe);
-        RecipeHolder<?> recipeHolder;
+        RecipeHolder<? extends IRecipe<?>> recipeHolder;
         switch (bukkitRecipe) {
             case ShapedRecipe shapedRecipe -> {
-                recipeHolder = ShapedRecipe260100.fromBukkit(recipeKey, shapedRecipe);
+                recipeHolder = ShapedRecipe12110.fromBukkit(recipeKey, shapedRecipe);
             }
             case ShapelessRecipe shapelessRecipe -> {
-                recipeHolder = ShapelessRecipe260100.fromBukkit(recipeKey, shapelessRecipe);
+                recipeHolder = ShapelessRecipe12110.fromBukkit(recipeKey, shapelessRecipe);
             }
             case FurnaceRecipe furnaceRecipe -> {
-                recipeHolder = SmeltingRecipe260100.fromBukkit(recipeKey, furnaceRecipe);
+                recipeHolder = FurnaceRecipe12110.fromBukkit(recipeKey, furnaceRecipe);
             }
             case SmokingRecipe smokingRecipe -> {
-                recipeHolder = SmokingRecipe260100.fromBukkit(recipeKey, smokingRecipe);
+                recipeHolder = SmokingRecipe12110.fromBukkit(recipeKey, smokingRecipe);
             }
             case CampfireRecipe campfireRecipe -> {
-                recipeHolder = CampfireCookingRecipe260100.fromBukkit(recipeKey, campfireRecipe);
+                recipeHolder = CampfireRecipe12110.fromBukkit(recipeKey, campfireRecipe);
             }
             case BlastingRecipe blastingRecipe -> {
-                recipeHolder = BlastingRecipe260100.fromBukkit(recipeKey, blastingRecipe);
+                recipeHolder = BlastingRecipe12110.fromBukkit(recipeKey, blastingRecipe);
             }
             case SmithingTransformRecipe smithingTransformRecipe -> {
-                recipeHolder = SmithingTransformRecipe260100.fromBukkit(recipeKey, smithingTransformRecipe);
+                recipeHolder = SmithingTransformRecipe12110.fromBukkit(recipeKey, smithingTransformRecipe);
             }
             case SmithingTrimRecipe smithingTrimRecipe -> {
-                recipeHolder = SmithingTrimRecipe260100.fromBukkit(recipeKey, smithingTrimRecipe);
+                recipeHolder = SmithingTrimRecipe12110.fromBukkit(recipeKey, smithingTrimRecipe);
             }
             case StonecuttingRecipe stonecuttingRecipe -> {
                 stonecuttingRecipe.setInputChoice(RecipeUtils.getBukkitChoice(stonecuttingRecipe.getInputChoice()));
@@ -73,14 +69,16 @@ public enum NmsRecipeRegister260100 implements NmsRecipeRegister, LifeCycleTask 
                 return RegisterResult.UNSUPPORTED_RECIPE_TYPE;
             }
         }
-        MinecraftServer.getServer().getRecipeManager().addRecipe(recipeHolder);
+        MinecraftServer.getServer().aN().addRecipe(recipeHolder);
         return RegisterResult.SUCCESS;
     }
 
     @Override
     public void lifecycle(Object plugin, LifeCycle lifeCycle) {
         if (!CrypticLibBukkit.isPaper()) {
-            NMS_REGISTER_COMPAT.register(MinecraftVersion.V26_1.name(), () -> this);
+            REGISTER_COMPAT.register(MinecraftVersion.V1_21_9.name(), () -> this);
         }
     }
 }
+
+
