+++ nms/v1_21_11_spigot/src/main/java/pers/yufiria/craftorithm/recipe/nms/spigot/CraftorithmRecipeRegister12111.java
@@ -1,5 +1,6 @@
-package pers.yufiria.craftorithm.recipe.nms;
+package pers.yufiria.craftorithm.recipe.nms.spigot;
 
+import crypticlib.CrypticLibBukkit;
 import crypticlib.MinecraftVersion;
 import crypticlib.lifecycle.LifeCycle;
 import crypticlib.lifecycle.LifeCycleTask;
@@ -7,24 +8,22 @@ import crypticlib.lifecycle.LifeCycleTaskSettings;
 import crypticlib.lifecycle.TaskRule;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.item.crafting.*;
+import org.bukkit.Bukkit;
 import org.bukkit.NamespacedKey;
-import org.bukkit.craftbukkit.v1_21_R5.inventory.CraftStonecuttingRecipe;
-import org.bukkit.craftbukkit.v1_21_R5.inventory.CraftTransmuteRecipe;
+import org.bukkit.craftbukkit.v1_21_R7.inventory.CraftStonecuttingRecipe;
 import org.bukkit.inventory.*;
 import org.bukkit.inventory.FurnaceRecipe;
 import org.bukkit.inventory.SmithingTransformRecipe;
 import org.bukkit.inventory.SmithingTrimRecipe;
 import org.bukkit.inventory.TransmuteRecipe;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
+import pers.yufiria.craftorithm.api.recipe.CraftorithmRecipeRegister;
 import pers.yufiria.craftorithm.recipe.RecipeManager;
 import pers.yufiria.craftorithm.util.RecipeUtils;
 
-@LifeCycleTaskSettings(
-    rules = {
-        @TaskRule(lifeCycle = LifeCycle.LOAD)
-    }
-)
-public enum NmsRecipeRegister12107 implements NmsRecipeRegister, LifeCycleTask {
+@LifeCycleTaskSettings(rules = {
+    @TaskRule(lifeCycle = LifeCycle.LOAD)
+})
+public enum CraftorithmRecipeRegister12111 implements CraftorithmRecipeRegister, LifeCycleTask {
 
     INSTANCE;
 
@@ -34,53 +33,52 @@ public enum NmsRecipeRegister12107 implements NmsRecipeRegister, LifeCycleTask {
         RecipeHolder<? extends IRecipe<?>> recipeHolder;
         switch (bukkitRecipe) {
             case ShapedRecipe shapedRecipe -> {
-                recipeHolder = ShapedRecipe12107.fromBukkit(recipeKey, shapedRecipe);
+                recipeHolder = ShapedRecipe12111.fromBukkit(recipeKey, shapedRecipe);
             }
             case ShapelessRecipe shapelessRecipe -> {
-                recipeHolder = ShapelessRecipe12107.fromBukkit(recipeKey, shapelessRecipe);
+                recipeHolder = ShapelessRecipe12111.fromBukkit(recipeKey, shapelessRecipe);
             }
             case FurnaceRecipe furnaceRecipe -> {
-                recipeHolder = FurnaceRecipe12107.fromBukkit(recipeKey, furnaceRecipe);
+                recipeHolder = FurnaceRecipe12111.fromBukkit(recipeKey, furnaceRecipe);
             }
             case SmokingRecipe smokingRecipe -> {
-                recipeHolder = SmokingRecipe12107.fromBukkit(recipeKey, smokingRecipe);
+                recipeHolder = SmokingRecipe12111.fromBukkit(recipeKey, smokingRecipe);
             }
             case CampfireRecipe campfireRecipe -> {
-                recipeHolder = CampfireRecipe12107.fromBukkit(recipeKey, campfireRecipe);
+                recipeHolder = CampfireRecipe12111.fromBukkit(recipeKey, campfireRecipe);
             }
             case BlastingRecipe blastingRecipe -> {
-                recipeHolder = BlastingRecipe12107.fromBukkit(recipeKey, blastingRecipe);
+                recipeHolder = BlastingRecipe12111.fromBukkit(recipeKey, blastingRecipe);
             }
             case SmithingTransformRecipe smithingTransformRecipe -> {
-                recipeHolder = SmithingTransformRecipe12107.fromBukkit(recipeKey, smithingTransformRecipe);
+                recipeHolder = SmithingTransformRecipe12111.fromBukkit(recipeKey, smithingTransformRecipe);
             }
             case SmithingTrimRecipe smithingTrimRecipe -> {
-                recipeHolder = SmithingTrimRecipe12107.fromBukkit(recipeKey, smithingTrimRecipe);
+                recipeHolder = SmithingTrimRecipe12111.fromBukkit(recipeKey, smithingTrimRecipe);
             }
             case StonecuttingRecipe stonecuttingRecipe -> {
-                //切石机配方自定义匹配逻辑存在问题，暂时还是按照原版的来
                 stonecuttingRecipe.setInputChoice(RecipeUtils.getBukkitChoice(stonecuttingRecipe.getInputChoice()));
                 CraftStonecuttingRecipe.fromBukkitRecipe(stonecuttingRecipe).addToCraftingManager();
                 return RegisterResult.SUCCESS;
             }
             case TransmuteRecipe transmuteRecipe -> {
-                CraftTransmuteRecipe.fromBukkitRecipe(transmuteRecipe).addToCraftingManager();
+                Bukkit.addRecipe(transmuteRecipe);
                 return RegisterResult.SUCCESS;
             }
             default -> {
                 return RegisterResult.UNSUPPORTED_RECIPE_TYPE;
             }
         }
-        MinecraftServer.getServer().aI().addRecipe(recipeHolder);
+        MinecraftServer.getServer().aJ().addRecipe(recipeHolder);
         return RegisterResult.SUCCESS;
     }
 
     @Override
     public void lifecycle(Object plugin, LifeCycle lifeCycle) {
-        NMS_REGISTER_COMPAT.register(
-            MinecraftVersion.V1_21_7.name(),
-            () -> this
-        );
+        if (!CrypticLibBukkit.isPaper()) {
+            REGISTER_COMPAT.register(MinecraftVersion.V1_21_11.name(), () -> this);
+        }
     }
-
 }
+
+
