+++ nms/v1_21_3/src/main/java/pers/yufiria/craftorithm/recipe/nms/CraftorithmRecipeRegister12103.java
@@ -1,6 +1,5 @@
-package pers.yufiria.craftorithm.recipe.nms.spigot;
+package pers.yufiria.craftorithm.recipe.nms;
 
-import crypticlib.CrypticLibBukkit;
 import crypticlib.MinecraftVersion;
 import crypticlib.lifecycle.LifeCycle;
 import crypticlib.lifecycle.LifeCycleTask;
@@ -10,20 +9,20 @@ import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.item.crafting.*;
 import org.bukkit.Bukkit;
 import org.bukkit.NamespacedKey;
-import org.bukkit.craftbukkit.v1_21_R6.inventory.CraftStonecuttingRecipe;
+import org.bukkit.craftbukkit.v1_21_R2.inventory.CraftStonecuttingRecipe;
 import org.bukkit.inventory.*;
 import org.bukkit.inventory.FurnaceRecipe;
 import org.bukkit.inventory.SmithingTransformRecipe;
 import org.bukkit.inventory.SmithingTrimRecipe;
 import org.bukkit.inventory.TransmuteRecipe;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
+import pers.yufiria.craftorithm.api.recipe.CraftorithmRecipeRegister;
 import pers.yufiria.craftorithm.recipe.RecipeManager;
 import pers.yufiria.craftorithm.util.RecipeUtils;
 
 @LifeCycleTaskSettings(rules = {
     @TaskRule(lifeCycle = LifeCycle.LOAD)
 })
-public enum NmsRecipeRegister12110 implements NmsRecipeRegister, LifeCycleTask {
+public enum CraftorithmRecipeRegister12103 implements CraftorithmRecipeRegister, LifeCycleTask {
 
     INSTANCE;
 
@@ -33,28 +32,28 @@ public enum NmsRecipeRegister12110 implements NmsRecipeRegister, LifeCycleTask {
         RecipeHolder<? extends IRecipe<?>> recipeHolder;
         switch (bukkitRecipe) {
             case ShapedRecipe shapedRecipe -> {
-                recipeHolder = ShapedRecipe12110.fromBukkit(recipeKey, shapedRecipe);
+                recipeHolder = ShapedRecipe12103.fromBukkit(recipeKey, shapedRecipe);
             }
             case ShapelessRecipe shapelessRecipe -> {
-                recipeHolder = ShapelessRecipe12110.fromBukkit(recipeKey, shapelessRecipe);
+                recipeHolder = ShapelessRecipe12103.fromBukkit(recipeKey, shapelessRecipe);
             }
             case FurnaceRecipe furnaceRecipe -> {
-                recipeHolder = FurnaceRecipe12110.fromBukkit(recipeKey, furnaceRecipe);
+                recipeHolder = FurnaceRecipe12103.fromBukkit(recipeKey, furnaceRecipe);
             }
             case SmokingRecipe smokingRecipe -> {
-                recipeHolder = SmokingRecipe12110.fromBukkit(recipeKey, smokingRecipe);
+                recipeHolder = SmokingRecipe12103.fromBukkit(recipeKey, smokingRecipe);
             }
             case CampfireRecipe campfireRecipe -> {
-                recipeHolder = CampfireRecipe12110.fromBukkit(recipeKey, campfireRecipe);
+                recipeHolder = CampfireRecipe12103.fromBukkit(recipeKey, campfireRecipe);
             }
             case BlastingRecipe blastingRecipe -> {
-                recipeHolder = BlastingRecipe12110.fromBukkit(recipeKey, blastingRecipe);
+                recipeHolder = BlastingRecipe12103.fromBukkit(recipeKey, blastingRecipe);
             }
             case SmithingTransformRecipe smithingTransformRecipe -> {
-                recipeHolder = SmithingTransformRecipe12110.fromBukkit(recipeKey, smithingTransformRecipe);
+                recipeHolder = SmithingTransformRecipe12103.fromBukkit(recipeKey, smithingTransformRecipe);
             }
             case SmithingTrimRecipe smithingTrimRecipe -> {
-                recipeHolder = SmithingTrimRecipe12110.fromBukkit(recipeKey, smithingTrimRecipe);
+                recipeHolder = SmithingTrimRecipe12103.fromBukkit(recipeKey, smithingTrimRecipe);
             }
             case StonecuttingRecipe stonecuttingRecipe -> {
                 stonecuttingRecipe.setInputChoice(RecipeUtils.getBukkitChoice(stonecuttingRecipe.getInputChoice()));
@@ -69,16 +68,12 @@ public enum NmsRecipeRegister12110 implements NmsRecipeRegister, LifeCycleTask {
                 return RegisterResult.UNSUPPORTED_RECIPE_TYPE;
             }
         }
-        MinecraftServer.getServer().aN().addRecipe(recipeHolder);
+        MinecraftServer.getServer().aI().addRecipe(recipeHolder);
         return RegisterResult.SUCCESS;
     }
 
     @Override
     public void lifecycle(Object plugin, LifeCycle lifeCycle) {
-        if (!CrypticLibBukkit.isPaper()) {
-            NMS_REGISTER_COMPAT.register(MinecraftVersion.V1_21_9.name(), () -> this);
-        }
+        REGISTER_COMPAT.register(MinecraftVersion.V1_21_3.name(), () -> this);
     }
 }
-
-
