+++ nms/v1_21_5/src/main/java/pers/yufiria/craftorithm/recipe/nms/CraftorithmRecipeRegister12105.java
@@ -6,67 +6,75 @@ import crypticlib.lifecycle.LifeCycleTask;
 import crypticlib.lifecycle.LifeCycleTaskSettings;
 import crypticlib.lifecycle.TaskRule;
 import net.minecraft.server.MinecraftServer;
-import net.minecraft.world.item.crafting.IRecipe;
-import net.minecraft.world.item.crafting.RecipeHolder;
+import net.minecraft.world.item.crafting.*;
+import org.bukkit.Bukkit;
 import org.bukkit.NamespacedKey;
-import org.bukkit.craftbukkit.v1_20_R4.inventory.CraftStonecuttingRecipe;
+import org.bukkit.craftbukkit.v1_21_R4.inventory.CraftStonecuttingRecipe;
 import org.bukkit.inventory.*;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
+import org.bukkit.inventory.FurnaceRecipe;
+import org.bukkit.inventory.SmithingTransformRecipe;
+import org.bukkit.inventory.SmithingTrimRecipe;
+import org.bukkit.inventory.TransmuteRecipe;
+import pers.yufiria.craftorithm.api.recipe.CraftorithmRecipeRegister;
 import pers.yufiria.craftorithm.recipe.RecipeManager;
 import pers.yufiria.craftorithm.util.RecipeUtils;
 
 @LifeCycleTaskSettings(rules = {
     @TaskRule(lifeCycle = LifeCycle.LOAD)
 })
-public enum NmsRecipeRegister12005 implements NmsRecipeRegister, LifeCycleTask {
+public enum CraftorithmRecipeRegister12105 implements CraftorithmRecipeRegister, LifeCycleTask {
 
     INSTANCE;
 
     @Override
     public RegisterResult registerRecipe(Recipe bukkitRecipe) {
         NamespacedKey recipeKey = RecipeManager.INSTANCE.getRecipeKey(bukkitRecipe);
-        RecipeHolder<? extends IRecipe<?>> nms;
+        RecipeHolder<? extends IRecipe<?>> recipeHolder;
         switch (bukkitRecipe) {
             case ShapedRecipe shapedRecipe -> {
-                nms = ShapedRecipe12005.fromBukkit(recipeKey, shapedRecipe);
+                recipeHolder = ShapedRecipe12105.fromBukkit(recipeKey, shapedRecipe);
             }
             case ShapelessRecipe shapelessRecipe -> {
-                nms = ShapelessRecipe12005.fromBukkit(recipeKey, shapelessRecipe);
+                recipeHolder = ShapelessRecipe12105.fromBukkit(recipeKey, shapelessRecipe);
             }
             case FurnaceRecipe furnaceRecipe -> {
-                nms = FurnaceRecipe12005.fromBukkit(recipeKey, furnaceRecipe);
+                recipeHolder = FurnaceRecipe12105.fromBukkit(recipeKey, furnaceRecipe);
             }
             case SmokingRecipe smokingRecipe -> {
-                nms = SmokingRecipe12005.fromBukkit(recipeKey, smokingRecipe);
+                recipeHolder = SmokingRecipe12105.fromBukkit(recipeKey, smokingRecipe);
             }
             case CampfireRecipe campfireRecipe -> {
-                nms = CampfireRecipe12005.fromBukkit(recipeKey, campfireRecipe);
+                recipeHolder = CampfireRecipe12105.fromBukkit(recipeKey, campfireRecipe);
             }
             case BlastingRecipe blastingRecipe -> {
-                nms = BlastingRecipe12005.fromBukkit(recipeKey, blastingRecipe);
+                recipeHolder = BlastingRecipe12105.fromBukkit(recipeKey, blastingRecipe);
             }
             case SmithingTransformRecipe smithingTransformRecipe -> {
-                nms = SmithingTransformRecipe12005.fromBukkit(recipeKey, smithingTransformRecipe);
+                recipeHolder = SmithingTransformRecipe12105.fromBukkit(recipeKey, smithingTransformRecipe);
             }
             case SmithingTrimRecipe smithingTrimRecipe -> {
-                nms = SmithingTrimRecipe12005.fromBukkit(recipeKey, smithingTrimRecipe);
+                recipeHolder = SmithingTrimRecipe12105.fromBukkit(recipeKey, smithingTrimRecipe);
             }
             case StonecuttingRecipe stonecuttingRecipe -> {
-                //切石机配方自定义匹配逻辑存在问题，暂时还是按照原版的来
                 stonecuttingRecipe.setInputChoice(RecipeUtils.getBukkitChoice(stonecuttingRecipe.getInputChoice()));
                 CraftStonecuttingRecipe.fromBukkitRecipe(stonecuttingRecipe).addToCraftingManager();
+                return RegisterResult.SUCCESS;
+            }
+            case TransmuteRecipe transmuteRecipe -> {
+                Bukkit.addRecipe(transmuteRecipe);
                 return RegisterResult.SUCCESS;
             }
             default -> {
                 return RegisterResult.UNSUPPORTED_RECIPE_TYPE;
             }
         }
-        MinecraftServer.getServer().aJ().addRecipe(nms);
+        MinecraftServer.getServer().aI().addRecipe(recipeHolder);
         return RegisterResult.SUCCESS;
     }
 
     @Override
-    public void lifecycle(Object o, LifeCycle lifeCycle) {
-        NMS_REGISTER_COMPAT.register(MinecraftVersion.V1_20_5.name(), () -> this);
+    public void lifecycle(Object plugin, LifeCycle lifeCycle) {
+        REGISTER_COMPAT.register(MinecraftVersion.V1_21_5.name(), () -> this);
     }
 }
+
