+++ nms/v1_21_7/src/main/java/pers/yufiria/craftorithm/recipe/nms/CraftorithmRecipeRegister12107.java
@@ -1,6 +1,5 @@
-package pers.yufiria.craftorithm.recipe.nms.paper;
+package pers.yufiria.craftorithm.recipe.nms;
 
-import crypticlib.CrypticLibBukkit;
 import crypticlib.MinecraftVersion;
 import crypticlib.lifecycle.LifeCycle;
 import crypticlib.lifecycle.LifeCycleTask;
@@ -8,80 +7,80 @@ import crypticlib.lifecycle.LifeCycleTaskSettings;
 import crypticlib.lifecycle.TaskRule;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.item.crafting.*;
-import org.bukkit.Bukkit;
 import org.bukkit.NamespacedKey;
-import org.bukkit.craftbukkit.inventory.CraftStonecuttingRecipe;
+import org.bukkit.craftbukkit.v1_21_R5.inventory.CraftStonecuttingRecipe;
+import org.bukkit.craftbukkit.v1_21_R5.inventory.CraftTransmuteRecipe;
 import org.bukkit.inventory.*;
-import org.bukkit.inventory.BlastingRecipe;
-import org.bukkit.inventory.Recipe;
-import org.bukkit.inventory.ShapedRecipe;
-import org.bukkit.inventory.ShapelessRecipe;
+import org.bukkit.inventory.FurnaceRecipe;
 import org.bukkit.inventory.SmithingTransformRecipe;
 import org.bukkit.inventory.SmithingTrimRecipe;
-import org.bukkit.inventory.SmokingRecipe;
 import org.bukkit.inventory.TransmuteRecipe;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
+import pers.yufiria.craftorithm.api.recipe.CraftorithmRecipeRegister;
 import pers.yufiria.craftorithm.recipe.RecipeManager;
 import pers.yufiria.craftorithm.util.RecipeUtils;
 
-@LifeCycleTaskSettings(rules = {
-    @TaskRule(lifeCycle = LifeCycle.LOAD)
-})
-public enum NmsRecipeRegister12110 implements NmsRecipeRegister, LifeCycleTask {
+@LifeCycleTaskSettings(
+    rules = {
+        @TaskRule(lifeCycle = LifeCycle.LOAD)
+    }
+)
+public enum CraftorithmRecipeRegister12107 implements CraftorithmRecipeRegister, LifeCycleTask {
 
     INSTANCE;
 
     @Override
     public RegisterResult registerRecipe(Recipe bukkitRecipe) {
         NamespacedKey recipeKey = RecipeManager.INSTANCE.getRecipeKey(bukkitRecipe);
-        RecipeHolder<?> recipeHolder;
+        RecipeHolder<? extends IRecipe<?>> recipeHolder;
         switch (bukkitRecipe) {
             case ShapedRecipe shapedRecipe -> {
-                recipeHolder = ShapedRecipe12110.fromBukkit(recipeKey, shapedRecipe);
+                recipeHolder = ShapedRecipe12107.fromBukkit(recipeKey, shapedRecipe);
             }
             case ShapelessRecipe shapelessRecipe -> {
-                recipeHolder = ShapelessRecipe12110.fromBukkit(recipeKey, shapelessRecipe);
+                recipeHolder = ShapelessRecipe12107.fromBukkit(recipeKey, shapelessRecipe);
             }
             case FurnaceRecipe furnaceRecipe -> {
-                recipeHolder = SmeltingRecipe12110.fromBukkit(recipeKey, furnaceRecipe);
+                recipeHolder = FurnaceRecipe12107.fromBukkit(recipeKey, furnaceRecipe);
             }
             case SmokingRecipe smokingRecipe -> {
-                recipeHolder = SmokingRecipe12110.fromBukkit(recipeKey, smokingRecipe);
+                recipeHolder = SmokingRecipe12107.fromBukkit(recipeKey, smokingRecipe);
             }
             case CampfireRecipe campfireRecipe -> {
-                recipeHolder = CampfireCookingRecipe12110.fromBukkit(recipeKey, campfireRecipe);
+                recipeHolder = CampfireRecipe12107.fromBukkit(recipeKey, campfireRecipe);
             }
             case BlastingRecipe blastingRecipe -> {
-                recipeHolder = BlastingRecipe12110.fromBukkit(recipeKey, blastingRecipe);
+                recipeHolder = BlastingRecipe12107.fromBukkit(recipeKey, blastingRecipe);
             }
             case SmithingTransformRecipe smithingTransformRecipe -> {
-                recipeHolder = SmithingTransformRecipe12110.fromBukkit(recipeKey, smithingTransformRecipe);
+                recipeHolder = SmithingTransformRecipe12107.fromBukkit(recipeKey, smithingTransformRecipe);
             }
             case SmithingTrimRecipe smithingTrimRecipe -> {
-                recipeHolder = SmithingTrimRecipe12110.fromBukkit(recipeKey, smithingTrimRecipe);
+                recipeHolder = SmithingTrimRecipe12107.fromBukkit(recipeKey, smithingTrimRecipe);
             }
             case StonecuttingRecipe stonecuttingRecipe -> {
+                //切石机配方自定义匹配逻辑存在问题，暂时还是按照原版的来
                 stonecuttingRecipe.setInputChoice(RecipeUtils.getBukkitChoice(stonecuttingRecipe.getInputChoice()));
                 CraftStonecuttingRecipe.fromBukkitRecipe(stonecuttingRecipe).addToCraftingManager();
                 return RegisterResult.SUCCESS;
             }
             case TransmuteRecipe transmuteRecipe -> {
-                Bukkit.addRecipe(transmuteRecipe);
+                CraftTransmuteRecipe.fromBukkitRecipe(transmuteRecipe).addToCraftingManager();
                 return RegisterResult.SUCCESS;
             }
             default -> {
                 return RegisterResult.UNSUPPORTED_RECIPE_TYPE;
             }
         }
-        MinecraftServer.getServer().getRecipeManager().addRecipe(recipeHolder);
+        MinecraftServer.getServer().aI().addRecipe(recipeHolder);
         return RegisterResult.SUCCESS;
     }
 
     @Override
     public void lifecycle(Object plugin, LifeCycle lifeCycle) {
-        //这个版本兼容1.21.9-1.21.11
-        if (CrypticLibBukkit.isPaper()) {
-            NMS_REGISTER_COMPAT.register(MinecraftVersion.V1_21_9.name(), () -> this);
-        }
+        REGISTER_COMPAT.register(
+            MinecraftVersion.V1_21_7.name(),
+            () -> this
+        );
     }
+
 }
