+++ nms/v26_1_paper/src/main/java/pers/yufiria/craftorithm/recipe/nms/paper/CraftorithmRecipeRegister260100.java
@@ -1,4 +1,4 @@
-package pers.yufiria.craftorithm.recipe.nms.spigot;
+package pers.yufiria.craftorithm.recipe.nms.paper;
 
 import crypticlib.CrypticLibBukkit;
 import crypticlib.MinecraftVersion;
@@ -10,55 +10,59 @@ import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.item.crafting.*;
 import org.bukkit.Bukkit;
 import org.bukkit.NamespacedKey;
-import org.bukkit.craftbukkit.v1_21_R7.inventory.CraftStonecuttingRecipe;
+import org.bukkit.craftbukkit.inventory.CraftStonecuttingRecipe;
 import org.bukkit.inventory.*;
-import org.bukkit.inventory.FurnaceRecipe;
+import org.bukkit.inventory.BlastingRecipe;
+import org.bukkit.inventory.Recipe;
+import org.bukkit.inventory.ShapedRecipe;
+import org.bukkit.inventory.ShapelessRecipe;
 import org.bukkit.inventory.SmithingTransformRecipe;
 import org.bukkit.inventory.SmithingTrimRecipe;
+import org.bukkit.inventory.SmokingRecipe;
 import org.bukkit.inventory.TransmuteRecipe;
-import pers.yufiria.craftorithm.api.recipe.NmsRecipeRegister;
+import pers.yufiria.craftorithm.api.recipe.CraftorithmRecipeRegister;
 import pers.yufiria.craftorithm.recipe.RecipeManager;
 import pers.yufiria.craftorithm.util.RecipeUtils;
 
 @LifeCycleTaskSettings(rules = {
     @TaskRule(lifeCycle = LifeCycle.LOAD)
 })
-public enum NmsRecipeRegister12111 implements NmsRecipeRegister, LifeCycleTask {
+public enum CraftorithmRecipeRegister260100 implements CraftorithmRecipeRegister, LifeCycleTask {
 
     INSTANCE;
 
     @Override
     public RegisterResult registerRecipe(Recipe bukkitRecipe) {
         NamespacedKey recipeKey = RecipeManager.INSTANCE.getRecipeKey(bukkitRecipe);
-        RecipeHolder<? extends IRecipe<?>> recipeHolder;
+        RecipeHolder<?> recipeHolder;
         switch (bukkitRecipe) {
             case ShapedRecipe shapedRecipe -> {
-                recipeHolder = ShapedRecipe12111.fromBukkit(recipeKey, shapedRecipe);
+                recipeHolder = ShapedRecipe260100.fromBukkit(recipeKey, shapedRecipe);
             }
             case ShapelessRecipe shapelessRecipe -> {
-                recipeHolder = ShapelessRecipe12111.fromBukkit(recipeKey, shapelessRecipe);
+                recipeHolder = ShapelessRecipe260100.fromBukkit(recipeKey, shapelessRecipe);
             }
             case FurnaceRecipe furnaceRecipe -> {
-                recipeHolder = FurnaceRecipe12111.fromBukkit(recipeKey, furnaceRecipe);
+                recipeHolder = SmeltingRecipe260100.fromBukkit(recipeKey, furnaceRecipe);
             }
             case SmokingRecipe smokingRecipe -> {
-                recipeHolder = SmokingRecipe12111.fromBukkit(recipeKey, smokingRecipe);
+                recipeHolder = SmokingRecipe260100.fromBukkit(recipeKey, smokingRecipe);
             }
             case CampfireRecipe campfireRecipe -> {
-                recipeHolder = CampfireRecipe12111.fromBukkit(recipeKey, campfireRecipe);
+                recipeHolder = CampfireCookingRecipe260100.fromBukkit(recipeKey, campfireRecipe);
             }
             case BlastingRecipe blastingRecipe -> {
-                recipeHolder = BlastingRecipe12111.fromBukkit(recipeKey, blastingRecipe);
+                recipeHolder = BlastingRecipe260100.fromBukkit(recipeKey, blastingRecipe);
             }
             case SmithingTransformRecipe smithingTransformRecipe -> {
-                recipeHolder = SmithingTransformRecipe12111.fromBukkit(recipeKey, smithingTransformRecipe);
+                recipeHolder = SmithingTransformRecipe260100.fromBukkit(recipeKey, smithingTransformRecipe);
             }
             case SmithingTrimRecipe smithingTrimRecipe -> {
-                recipeHolder = SmithingTrimRecipe12111.fromBukkit(recipeKey, smithingTrimRecipe);
+                recipeHolder = SmithingTrimRecipe260100.fromBukkit(recipeKey, smithingTrimRecipe);
             }
             case StonecuttingRecipe stonecuttingRecipe -> {
                 stonecuttingRecipe.setInputChoice(RecipeUtils.getBukkitChoice(stonecuttingRecipe.getInputChoice()));
-                CraftStonecuttingRecipe.fromBukkitRecipe(stonecuttingRecipe).addToCraftingManager();
+                CraftStonecuttingRecipe.fromBukkitRecipe(stonecuttingRecipe).addToRecipeManager();
                 return RegisterResult.SUCCESS;
             }
             case TransmuteRecipe transmuteRecipe -> {
@@ -69,16 +73,14 @@ public enum NmsRecipeRegister12111 implements NmsRecipeRegister, LifeCycleTask {
                 return RegisterResult.UNSUPPORTED_RECIPE_TYPE;
             }
         }
-        MinecraftServer.getServer().aJ().addRecipe(recipeHolder);
+        MinecraftServer.getServer().getRecipeManager().addRecipe(recipeHolder);
         return RegisterResult.SUCCESS;
     }
 
     @Override
     public void lifecycle(Object plugin, LifeCycle lifeCycle) {
-        if (!CrypticLibBukkit.isPaper()) {
-            NMS_REGISTER_COMPAT.register(MinecraftVersion.V1_21_11.name(), () -> this);
+        if (CrypticLibBukkit.isPaper()) {
+            REGISTER_COMPAT.register(MinecraftVersion.V26_1.name(), () -> this);
         }
     }
 }
-
-
