--- src/main/resources/config.yml
@@ -1,55 +1,0 @@
-#插件的配置版本
-config_version: 2
-#是否开启版本更新检查
-check_update: true
-#是否移除所有的原版配方
-remove_all_vanilla_recipe: false
-#是否启用插件铁砧配方
-enable_anvil_recipe: true
-#是否启用bstats插件使用数据统计
-bstats: true
-#是否在ItemsAdder重载时也重载Craftorithm
-reload_when_ia_reload: true
-#是否开启调试
-debug: false
-#单个tick最多注册几个配方
-max_reg_recipe_per_tick: 20
-#是否使用NMS方式注册配方
-#通过nms注册的配方，在配方书等情况下可能会出现一些问题
-use_nms_recipe_register: false
-cannot_craft_items: []
-#依照上面的挂钩顺序挂钩插件可以挂钩的物品插件,插件自动识别物品ID时将会优先识别上面的插件
-#不包含在此列表里的物品插件将不会尝试挂钩,除非该插件主动挂钩
-item_plugin_hook_priority:
-  - CraftEngine
-  - Nexo
-  - AzureFlow
-  - NeigeItems
-  - ItemsAdder
-  - Oraxen
-  - EcoItems
-  - ExecutableItems
-  - MMOItems
-  - MythicMobs
-main_command_aliases:
-  - cra
-  - craft
-  - crafto
-not_convert_listener_classes:
-  - a4.papers.chatfilter.chatfilter.events.AnvilListener
-  - com.ghostchu.quickshop.shade.tne.menu.paper.listener.PaperInventoryClickListener
-  - com.earth2me.essentials.EssentialsPlayerListener
-  - net.coreprotect.listener.player.InventoryChangeListener
-  - net.coreprotect.listener.player.CraftItemListener
-  - com.extendedclip.deluxemenus.listener.PlayerListener
-  - com.dre.brewery.listeners.InventoryListener
-  - com.xyrisdev.svalues.shaded.library.menu.MenuManager$InventoryListener
-  - me.arcaniax.hdb.listener.InventoryListener
-  - net.momirealms.craftengine.bukkit.item.listener.ItemEventListener
-  - net.momirealms.customfishing.bukkit.hook.BukkitHookManager
-  - net.momirealms.customfishing.bukkit.market.BukkitMarketManager
-  - dev.jsinco.recipes.listeners.Events
-  - fr.moribus.imageonmap.image.MapInitEvent
-  - com.badbones69.crazycrates.paper.listeners.crates.types.WarCrateListener
-  - com.ryderbelserion.fusion.paper.api.builders.gui.listeners.GuiListener
-  - club.kid7.bannermaker.pluginutilities.gui.CustomGUIInventoryListener
